create files for layout , finish logic in moderator and admin controllers , create nessisary routes and test all functionality with blade , also update some migration files and controllers , next : manage images and start on front end'

+++ app/Http/Controllers/CategoryController.php
@@ -10,13 +10,21 @@ use Illuminate\Support\Facades\Auth;
 class CategoryController extends Controller
 {
 
+    public function index(){
+        $categories = Category::where('user_id' , Auth::id())->get() ;
+        return view('tasks.categories.index' , compact('categories'));
+
+    }
+
     public function create()
     {
+        $this->authorize('CreateGlobalCategory', Category::class);
         return view('tasks.categories.create');
     }
 
     public function store(StoreCategoryRequest $request)
     {
+        $this->authorize('CreateGlobalCategory', Category::class);
         $data = $request->validated();
         $data['user_id'] = Auth::id();
         $category = Category::create($data);
@@ -44,22 +52,28 @@ class CategoryController extends Controller
 
     public function edit(Category $category)
     {
+        $this->authorize('UpdateGlobalCategory', $category);
+        $this->authorize('update', $category);
         return view('tasks.categories.edit', compact('category'));
     }
 
 
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+        $this->authorize('UpdateGlobalCategory', $category);
+        $this->authorize('update', $category);
         $data = $request->validated();
         $category->update($data);
         $category->save();
         return redirect()->route('categories.show', $category);
     }
 
-
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
         $category->delete();
         return redirect()->route('categorie.index');
     }
+    
+
 }
