+++ app/Http/Controllers/HabitController.php
@@ -7,10 +7,12 @@ use App\Http\Requests\UpdateHabitRequest;
 
 use App\Models\Category;
 use App\Models\Task;
+use App\Policies\TaskPolicy;
 use Illuminate\Support\Facades\Auth;
 
 class HabitController extends Controller
 {
+
 
     public function index()
     {
@@ -46,22 +48,25 @@ class HabitController extends Controller
 
     public function edit(Task $habit)
     {
+        $this->authorize('update', $habit);
         $user = Auth::user();
         $categories = Category::where('user_id', $user->id)->get();
         return view('tasks.habits.edit', compact('habit', 'categories'));
     }
 
-    public function update(UpdateHabitRequest $request, Task $habit)
+    public function update(UpdateHabitRequest $request, Task $task)
     {
+        $this->authorize('update', $task);
         $newHabit = $request->validated();
-        $habit->update($newHabit);
-        $habit->save();
-        return redirect()->route('habits.show' ,$habit);
+        $task->update($newHabit);
+        $task->save();
+        return redirect()->route('habits.show' ,$task);
     }
 
 
     public function destroy(Task $habit)
     {
+        $this->authorize('delete', $habit);
         $habit->delete();
         return redirect()->route('habits.index') ;
     }
